+++ src/modules/genres/genre.entity.ts
@@ -25,5 +25,5 @@ export class Genre extends DatabaseEntity {
     type: () => Game,
     isArray: true,
   })
-  games?: Game[];
+  games: Game[];
 }