+++ src/modules/images/image.entity.ts
@@ -1,4 +1,4 @@
-import { ApiProperty } from "@nestjs/swagger";
+import { ApiPropertyOptional } from "@nestjs/swagger";
 import { Entity, Column, Index, ManyToOne } from "typeorm";
 import { DatabaseEntity } from "../database/database.entity";
 import { GamevaultUser } from "../users/gamevault-user.entity";
@@ -7,7 +7,7 @@ import { GamevaultUser } from "../users/gamevault-user.entity";
 export class Image extends DatabaseEntity {
   @Index()
   @Column({ nullable: true })
-  @ApiProperty({
+  @ApiPropertyOptional({
     example:
       "https://upload.wikimedia.org/wikipedia/commons/thumb/e/e1/Grand_Theft_Auto_logo_series.svg",
     description: "the original source URL of the image",
@@ -17,14 +17,14 @@ export class Image extends DatabaseEntity {
 
   @Column({ unique: true, nullable: true })
   @Index()
-  @ApiProperty({
+  @ApiPropertyOptional({
     example: "/images/14",
     description: "the path of the image on the filesystem",
   })
   path?: string;
 
   @Column({ nullable: true })
-  @ApiProperty({
+  @ApiPropertyOptional({
     example: "image/jpeg",
     description: "the media type of the image on the filesystem",
   })
@@ -33,10 +33,9 @@ export class Image extends DatabaseEntity {
   @ManyToOne(() => GamevaultUser, (user) => user.uploaded_images, {
     nullable: true,
   })
-  @ApiProperty({
+  @ApiPropertyOptional({
     description: "the uploader of the image",
     type: () => GamevaultUser,
-    nullable: true,
   })
   uploader?: GamevaultUser;
 }